email_app/requirements.txt
openai==1.59.6
python-dotenv==1.0.1
streamlit==1.41.1